@@ -7,22 +7,20 @@ plugins {
 android {
     signingConfigs {
         release {
-            storeFile file('W:\\Xmapp\\htdocs\\projects\\newpizzayum1\\NewPizzaYum.jks')
-            storePassword 'NewPizzaYum@123'
-            keyPassword 'NewPizzaYum@123'
-            keyAlias 'key0'
+            storeFile file('H:\\AndroidProjectKeyStore--DoNotTouchOrRemove\\com_mtechyard_newpizzayum.jks')
+            storePassword 'com_mtechyard_newpizzayum.jks'
+            keyPassword 'com_mtechyard_newpizzayum.jks'
+            keyAlias 'key'
         }
     }
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "com.mtechyard.newpizzayum"
         minSdkVersion 22
-        targetSdkVersion 29
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
-
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
@@ -45,22 +43,19 @@ android {
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
     // GOOGLE LOCATION API
-    implementation 'com.google.android.gms:play-services-location:18.0.0'
-    implementation 'androidx.navigation:navigation-fragment:2.3.3'
-    implementation 'androidx.navigation:navigation-ui:2.3.3'
-    implementation 'com.google.firebase:firebase-inappmessaging-display:19.1.4'
+    implementation 'com.google.android.gms:play-services-location:21.0.0'
+    implementation 'androidx.navigation:navigation-fragment:2.5.3'
+    implementation 'androidx.navigation:navigation-ui:2.5.3'
+    implementation 'com.google.firebase:firebase-inappmessaging-display:20.2.0'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-
-
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     implementation platform('com.google.firebase:firebase-bom:26.5.0')
-
     implementation 'com.google.firebase:firebase-analytics'
 
     // Declare the dependency for the Firebase Authentication library
@@ -68,24 +63,24 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth'
 
     // NETWORKING DEPENDENCIES
-    implementation 'com.android.volley:volley:1.1.1'
+    implementation 'com.android.volley:volley:1.2.1'
 
     // CIRCLE IMAGE DEPENDENCIES
     implementation 'de.hdodenhof:circleimageview:3.1.0'
 
     // GSON
-    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'com.google.code.gson:gson:2.9.0'
 
     // pin view
     implementation 'com.chaos.view:pinview:1.4.4'
 
     def lottieVersion = "3.6.1"
     implementation "com.airbnb.android:lottie:$lottieVersion"
+
     // picasso for online images
     implementation 'com.squareup.picasso:picasso:2.71828'
-
-
     implementation 'com.paytm.appinvokesdk:appinvokesdk:1.5.4'
+
     // android sdk
-    implementation 'com.razorpay:checkout:1.6.6'
+    implementation 'com.razorpay:checkout:1.6.26'
 }
