@@ -6,7 +6,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 
-import com.mtechyard.newpizzayum.project_rec.TinyDB;
+import com.mtechyard.newpizzayum.app.TinyDB;
+import com.mtechyard.newpizzayum.app_src.AppDB;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -15,25 +16,18 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-//        startActivity(new Intent(MainActivity.this, home.class));
-//        finish();
-
-
         Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                TinyDB tinyDB = new TinyDB(MainActivity.this);
-                String userMobileNo = tinyDB.getUserMobileNo();
+        handler.postDelayed(() -> {
+            AppDB tinyDB = new AppDB(MainActivity.this);
+            String userMobileNo = tinyDB.getUserMobileNo();
 
-                if (userMobileNo.trim().isEmpty()) {
-                    startActivity(new Intent(MainActivity.this, user_auth.class));
-                    finish();
-                } else {
-                    startActivity(new Intent(MainActivity.this, home.class));
-                    finish();
-                }
+            if (userMobileNo.trim().isEmpty()) {
+                startActivity(new Intent(MainActivity.this, UserAuthActivity.class));
+                finish();
+            } else {
+                startActivity(new Intent(MainActivity.this, HomeActivity.class));
+                finish();
             }
-        }, 1000);
+        }, 5000);
     }
 }
