@@ -1,4 +1,4 @@
-package com.mtechyard.newpizzayum.project_rec;
+package com.mtechyard.newpizzayum.api;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
