@@ -1,11 +1,9 @@
-package com.mtechyard.newpizzayum.project_rec;
+package com.mtechyard.newpizzayum.api;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 public class UserOrderData {
-
-
 
     @SerializedName("orderName")
     @Expose
